Move logout to Navbar + add /navigate after logout + Fix map for campagincharacters
